requirements.txt
django-environ~=0.4.5
dj-database-url==0.5.0
psycopg2~=2.9.2